Fixed small mistake
git-svn-id: http://calder.sdml.cs.kent.edu/svn/srcDiffPurged/trunk/srcdiff/trunk@11415 f70b1346-9fe7-0310-9bf3-95a0de396e6b

+++ diff_trace/SAX2DiffTrace.hpp
@@ -73,7 +73,7 @@ class SAX2DiffTrace {
  public:
 
   // constructor
-  SAX2DiffTrace(int & options);
+  SAX2DiffTrace(long & options);
 
   // sax factory
   static xmlSAXHandler factory();
